apps/video-worker/pyproject.toml
[project]
name = "video-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0.0"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "openai-whisper (>=20240930,<20240931)",
    "webvtt-py (>=0.5.1,<0.6.0)",
    "rapidfuzz (>=3.13.0,<4.0.0)",
    "cupy-cuda12x (>=13.4.1,<14.0.0)",
    "ultralytics (>=8.3.116,<9.0.0)",
    "pypino (>=0.1.0,<0.2.0)",
    "ray (>=2.45.0,<3.0.0)",
    "kafka-python (>=2.2.4,<3.0.0)",
    "torch (>=2.7.0,<3.0.0)",
    "xformers (>=0.0.30,<0.0.31)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
